# backend/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Docker Web Desktop API",
    "description": "API documentation for Docker Web Desktop (auto-generated)",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/commands/execute": {
      "post": {
        "tags": [
          "Commands"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "command": {
                  "example": "any"
                },
                "source": {
                  "example": "any"
                },
                "wslDistro": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/commands/openconsole": {
      "post": {
        "tags": [
          "Commands"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "command": {
                  "example": "any"
                },
                "source": {
                  "example": "any"
                },
                "wslDistro": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/commands/info": {
      "get": {
        "tags": [
          "Commands"
        ],
        "description": "",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/commands/search-image": {
      "post": {
        "tags": [
          "Commands"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "q": {
                  "example": "any"
                },
                "sort": {
                  "example": "any"
                },
                "is_official": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "408": {
            "description": "Request Timeout"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/commands/search-image-tags": {
      "get": {
        "tags": [
          "Commands"
        ],
        "description": "",
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "408": {
            "description": "Request Timeout"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}